Creating global & default configurations on demand
The previous method was to create the local configuration files during
installation. This would not work, if the package was installed into a
virtualenv the first time and the first usage would cause the
application to complain about missing global/default config files.

The new method is just to check if the files exist whenever they are
referenced.