Fix: RcfileModifier handles missing file
The RcfileModifier subclasses (AddToBashrcPath, etc.) all had the same
bug when testing if the extension was installed: They were not testing
if the file content was empty or None. Fixed this issue by accepting a
new parameter to RcfileModifier.read_file() which defines the return
value if the file does not exist. This makes the following call short
and safe:

    'foo' in modifier.read_file('')